Add Probe warning
