+++ internal/cmd/kickstart/kickstart_test.go
@@ -9,12 +9,22 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+var kickstartMessage = `Kickstarted the "my-pkg" package with 3 objects.
+[WARN] Some kinds don't have availability probes defined:
+- Banana.fruits
+`
+
 func TestKickstart(t *testing.T) {
-	defer os.RemoveAll("my-pkg")
+	t.Parallel()
+	defer func() {
+		if err := os.RemoveAll("my-pkg"); err != nil {
+			panic(err)
+		}
+	}()
 
 	ctx := context.Background()
 	k := NewKickstarter(nil)
 	msg, err := k.KickStart(ctx, "my-pkg", []string{"testdata/all-the-objects.yaml"})
 	require.NoError(t, err)
-	assert.Equal(t, `Kickstarted the "my-pkg" package with 2 objects.`, msg)
+	assert.Equal(t, kickstartMessage, msg)
 }
