+++ internal/cmd/kickstart/probing.go
@@ -1,0 +1,128 @@
+package kickstart
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+
+	corev1alpha1 "package-operator.run/apis/core/v1alpha1"
+)
+
+// Determines probes required for the given Group Kind.
+func getProbe(gk schema.GroupKind) (corev1alpha1.ObjectSetProbe, bool) {
+	probes, ok := gkProbes[gk]
+	if !ok {
+		return corev1alpha1.ObjectSetProbe{}, false
+	}
+	return corev1alpha1.ObjectSetProbe{
+		Selector: corev1alpha1.ProbeSelector{
+			Kind: &corev1alpha1.PackageProbeKindSpec{
+				Group: gk.Group,
+				Kind:  gk.Kind,
+			},
+		},
+		Probes: probes,
+	}, true
+}
+
+var gkProbes = map[schema.GroupKind][]corev1alpha1.Probe{
+	{
+		Kind: "Deployment", Group: "apps",
+	}: {
+		availableProbe,
+		replicasUpdatedProbe,
+	},
+	{
+		Kind: "StatefulSet", Group: "apps",
+	}: {
+		availableProbe,
+		replicasUpdatedProbe,
+	},
+	{
+		Kind: "DaemonSet", Group: "apps",
+	}: {
+		{
+			FieldsEqual: &corev1alpha1.ProbeFieldsEqualSpec{
+				FieldA: ".status.desiredNumberScheduled",
+				FieldB: ".status.numberAvailable",
+			},
+		},
+	},
+	{
+		Kind: "ReplicaSet", Group: "apps",
+	}: {
+		availableProbe,
+		replicasUpdatedProbe,
+	},
+	{
+		Kind:  "CustomResourceDefinition",
+		Group: "apiextensions.k8s.io",
+	}: {
+		{
+			Condition: &corev1alpha1.ProbeConditionSpec{
+				Type:   "Established",
+				Status: string(metav1.ConditionTrue),
+			},
+		},
+	},
+	{
+		Kind:  "Job",
+		Group: "batch",
+	}: {
+		{
+			Condition: &corev1alpha1.ProbeConditionSpec{
+				Type:   "Complete",
+				Status: string(metav1.ConditionTrue),
+			},
+		},
+	},
+	{
+		Kind:  "Route",
+		Group: "route.openshift.io",
+	}: {
+		{
+			CEL: &corev1alpha1.ProbeCELSpec{
+				Message: "not all ingress points are reporting ready",
+				Rule:    `self.status.ingress.all(i, i.conditions.all(c, c.type == "Ready" && c.status == "True"))`,
+			},
+		},
+	},
+	{
+		Kind:  "PersistentVolumeClaim",
+		Group: "",
+	}: {
+		{
+			CEL: &corev1alpha1.ProbeCELSpec{
+				Message: "is not yet Bound",
+				Rule:    `self.status.phase == "Bound"`,
+			},
+		},
+	},
+	{
+		Kind:  "ClusterServiceVersion",
+		Group: "operators.coreos.com",
+	}: {
+		{
+			CEL: &corev1alpha1.ProbeCELSpec{
+				Message: "CSV not succeeded",
+				Rule:    `self.status.phase == "Succeeded"`,
+			},
+		},
+	},
+}
+
+// Checks if the Available Condition is True.
+var availableProbe = corev1alpha1.Probe{
+	Condition: &corev1alpha1.ProbeConditionSpec{
+		Type:   "Available",
+		Status: string(metav1.ConditionTrue),
+	},
+}
+
+// Checks if all replicas have been updated.
+// Works for StatefulSets, Deployments and ReplicaSets.
+var replicasUpdatedProbe = corev1alpha1.Probe{
+	FieldsEqual: &corev1alpha1.ProbeFieldsEqualSpec{
+		FieldA: ".status.updatedReplicas",
+		FieldB: ".status.replicas",
+	},
+}
