+++ internal/cmd/kickstart/testdata/all-the-objects.yaml
@@ -29,3 +29,10 @@ metadata:
     pod-security.kubernetes.io/audit-version: latest
     pod-security.kubernetes.io/warn: restricted
     pod-security.kubernetes.io/warn-version: latest
+---
+apiVersion: fruits/v1
+kind: Banana
+metadata:
+  name: cavendish
+spec:
+  sweet: True
